Edit ggplot() so that class of spct objects is not dropped.
ggplot2::ggplot.data.frame() cannot be called directly as it is private. So the code for a long time has converted the objects into tibbles or data frames for method dispatch to work. This update adds code to set the class attribute of the copy of the spectrum stored in the "gg" object to the original one. This should allow correct dispatch of methods when modifying data in layers. I also eases the task of retrieving metadata.