Fix bug in autoplot() methods
All autoplot methods could map to linetype more factor levels than the 13 patterns available. Even more than 9 or so patterns were difficult to distinguish in the too narrow keys. Because of a bug in 'photobiology' 0.10.15 autoplot() methods could crash with an error for collections of spectra containing only a single spct object with multiple spectra in long form.