Add fast_wb2rect_df()
fast_wb2rect_df() is an internal function intended to replace calls to photobiology::fast_wb2rect_spct() that only computes values used in the statistics and returns them in a data frame.
Remove the states.var argument recently added to stat_wl_strip.Rd as in principle annimatiion work automatically in the abscence of grouping.