Fix normalization label bug in all autoplot() methods
As 'photobiology' 0.14.0 supports normalization attribute for spectra in long form, its extraction for use in labels needs to handle lists or normalization records. The new code, not only solves the handling of this case, but generates the most appropriate label, based on wavelength or type of normalization depending on the case.